+++ GMC_Run/GMC_Run/running_stat.h
@@ -4,11 +4,11 @@ class RunningStat
 public:
     RunningStat();
     void Clear();
-    void Push(long double x);
+    void Push(float x);
     int NumDataValues();
-    long double Mean();
-    long double Variance();
+    float Mean();
+    float Variance();
 private:
     int m_n;
-    long double m_oldM, m_newM, m_oldS, m_newS;
+    float m_oldM, m_newM, m_oldS, m_newS;
 };
